requirements.txt
arrow==0.10.0
Django==1.10.6
django-disqus==0.5
django-forms-bootstrap==3.0.1
django-paypal==0.3.3
funcsigs==1.0.2
mock==2.0.0
olefile==0.44
pbr==2.0.0
Pillow==2.9.0
python-dateutil==2.6.0
pytz==2016.10
requests==2.13.0
six==1.10.0
wheel==0.24.0
git+https://github.com/meganduffy/reusable_blogprj_app.git
